+++ llama-cpp-bindings/src/error/llama_context_load_error.rs
@@ -7,5 +7,5 @@ pub enum LlamaContextLoadError {
     #[error("{message}")]
     Reported { message: String },
     #[error("the FFI wrapper returned an unrecognized status code {code}")]
-    UnrecognizedStatusCode { code: u32 },
+    UnrecognizedStatusCode { code: i64 },
 }
