+++ llama-cpp-bindings/src/error/sample_error.rs
@@ -14,5 +14,5 @@ pub enum SampleError {
     #[error("{message}")]
     Reported { message: String },
     #[error("the FFI wrapper returned an unrecognized status code {code}")]
-    UnrecognizedStatusCode { code: u32 },
+    UnrecognizedStatusCode { code: i64 },
 }
